# backend/docs/swagger.yaml
basePath: /
definitions:
  models.CDNLine:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
      provider:
        $ref: '#/definitions/models.CDNProvider'
      provider_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.CDNProvider:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateVideoStreamEndpointRequest:
    properties:
      domain_id:
        type: integer
      line_id:
        type: integer
      provider_id:
        type: integer
      status:
        type: integer
      stream_id:
        type: integer
      stream_path_id:
        type: integer
    required:
    - domain_id
    - line_id
    - provider_id
    - stream_id
    - stream_path_id
    type: object
  models.Domain:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Stream:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.StreamPath:
    properties:
      created_at:
        type: string
      full_path:
        type: string
      id:
        type: integer
      stream:
        $ref: '#/definitions/models.Stream'
      stream_id:
        type: integer
      table_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateVideoStreamEndpointRequest:
    properties:
      domain_id:
        type: integer
      line_id:
        type: integer
      provider_id:
        type: integer
      status:
        type: integer
      stream_id:
        type: integer
      stream_path_id:
        type: integer
    required:
    - domain_id
    - line_id
    - provider_id
    - stream_id
    - stream_path_id
    type: object
  models.VideoStreamEndpoint:
    properties:
      created_at:
        type: string
      domain:
        $ref: '#/definitions/models.Domain'
      domain_id:
        type: integer
      full_url:
        type: string
      id:
        type: integer
      line:
        $ref: '#/definitions/models.CDNLine'
      line_id:
        type: integer
      provider:
        $ref: '#/definitions/models.CDNProvider'
      provider_id:
        type: integer
      status:
        type: integer
      stream:
        $ref: '#/definitions/models.Stream'
      stream_id:
        type: integer
      stream_path:
        $ref: '#/definitions/models.StreamPath'
      stream_path_id:
        type: integer
      updated_at:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Video Manager API Documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Video Manager API
  version: "1.0"
paths:
...
  /api/video-stream-endpoints:
    get:
      description: Retrieve all video stream endpoints, optionally filtered by line_id,
        domain_id, stream_id, provider_id, status, or table_id
      parameters:
      - description: Filter by line ID
        in: query
        name: line_id
        type: integer
      - description: Filter by domain ID
        in: query
        name: domain_id
        type: integer
      - description: Filter by stream ID
        in: query
        name: stream_id
        type: integer
      - description: Filter by provider ID
        in: query
        name: provider_id
        type: integer
      - description: Filter by status (0=disabled, 1=enabled)
        in: query
        name: status
        type: integer
      - description: Filter by table ID (桌台号)
        in: query
        name: table_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.VideoStreamEndpoint'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all video stream endpoints
      tags:
      - video-stream-endpoints
    post:
      consumes:
      - application/json
      description: Create a new video stream endpoint with the provided information
      parameters:
      - description: Video Stream Endpoint information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateVideoStreamEndpointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoStreamEndpoint'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new video stream endpoint
      tags:
      - video-stream-endpoints
  /api/video-stream-endpoints/{id}:
    delete:
      description: Delete a video stream endpoint by its ID
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a video stream endpoint
      tags:
      - video-stream-endpoints
    get:
      description: Retrieve a video stream endpoint by its ID
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoStreamEndpoint'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get video stream endpoint by ID
      tags:
      - video-stream-endpoints
    put:
      consumes:
      - application/json
      description: Update an existing video stream endpoint by its ID
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated video stream endpoint information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateVideoStreamEndpointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoStreamEndpoint'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a video stream endpoint
      tags:
      - video-stream-endpoints
  /api/video-stream-endpoints/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status (enabled/disabled) of a video stream endpoint
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status (0=disabled, 1=enabled)
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update video stream endpoint status
      tags:
      - video-stream-endpoints
  /api/video-stream-endpoints/generate:
    post:
      description: Automatically generate all video stream endpoints based on all
        combinations of CDN lines, domains, and stream paths. This will delete all
        existing endpoints and regenerate them.
      produces:
      - application/json
      responses:
        "200":
          description: Response contains message and count
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Generate all video stream endpoints
      tags:
      - video-stream-endpoints
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"